requirements.txt
streamlit>=1.30.0
transformers>=4.35.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
gTTS>=2.4.0
